refactor: extract shared youtube player ad muter
